rename+pictures fixes
